requirements.txt
python-dotenv>=1.0.0
langgraph>=1.1.0
langchain-core>=1.0.0
langchain-deepseek>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
